+++ lib/table_sync/receiving.rb
@@ -2,6 +2,8 @@
 
 module TableSync
   module Receiving
+    AVAILABLE_EVENTS = [:update, :destroy].freeze
+
     require_relative "receiving/config"
     require_relative "receiving/config_decorator"
     require_relative "receiving/dsl"